CIV-9253 excluding model from sonar
