+++ src/main/java/uk/gov/hmcts/reform/civil/model/prd/client/OrganisationApi.java
@@ -1,10 +1,10 @@
-package uk.gov.hmcts.reform.civil.prd.client;
+package uk.gov.hmcts.reform.civil.model.prd.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
-import uk.gov.hmcts.reform.civil.prd.model.Organisation;
+import uk.gov.hmcts.reform.civil.model.prd.model.Organisation;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi.SERVICE_AUTHORIZATION;
