+++ src/main/java/uk/gov/hmcts/reform/civil/service/OrganisationService.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.civil.prd.client.OrganisationApi;
-import uk.gov.hmcts.reform.civil.prd.model.Organisation;
+import uk.gov.hmcts.reform.civil.model.prd.client.OrganisationApi;
+import uk.gov.hmcts.reform.civil.model.prd.model.Organisation;
 
 import java.util.Optional;
