+++ src/test/java/uk/gov/hmcts/reform/civil/service/OrganisationServiceTest.java
@@ -10,8 +10,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.civil.prd.client.OrganisationApi;
-import uk.gov.hmcts.reform.civil.prd.model.Organisation;
+import uk.gov.hmcts.reform.civil.model.prd.client.OrganisationApi;
+import uk.gov.hmcts.reform.civil.model.prd.model.Organisation;
 
 import java.util.Map;
 import java.util.Optional;
