+++ src/test/java/uk/gov/hmcts/reform/civil/service/SubmitCreateClaimServiceTest.java
@@ -15,7 +15,7 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.civil.config.CreateClaimConfiguration;
-import uk.gov.hmcts.reform.civil.prd.model.Organisation;
+import uk.gov.hmcts.reform.civil.model.prd.model.Organisation;
 import uk.gov.hmcts.reform.civil.requestbody.CreateClaimCCD;
 import uk.gov.hmcts.reform.civil.responsebody.CreateClaimResponse;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
